usuarios

+++ backend/src/controllers/eventController.js
@@ -3,7 +3,11 @@ import Event from "../models/event.js";
 
 const getEvents = async (req, res) => {
     try {
-        const events = await Event.find({});
+        let user = req.user;
+        if(!user) {
+            return res.status(401).json({message:"User not authenticated"});
+        }
+        const events = await Event.find({user:user.email});
         res.json(events);
     } catch (error) {
         res.status(500).json({message: error.message});
@@ -12,8 +16,13 @@ const getEvents = async (req, res) => {
 
 const addEvent = async (req, res) => {
     console.log("body ",req.body)
+    let user = req.user;
+    if(!user) {
+        return res.status(401).json({message:"User not authenticated"});
+    }
     const event = new Event({
-        event: req.body.event
+        event: req.body.event,
+        user: user.email
     });
     try {
         const newEvent = await event.save();
@@ -25,9 +34,16 @@ const addEvent = async (req, res) => {
 
 const deleteEvent = async (req, res) => {
     try {
+        let user = req.user;
+        if(!user) {
+            return res.status(401).json({message:"User not authenticated"});
+        }
         const event = await Event.findById(req.params._id);
         if (!event) {
             return res.status(404).json({message: 'Event not found'});
+        }
+        if (event.user !== user.email && !user.isAdmin) {
+            return res.status(401).json({message: 'User not authorized'});
         }
         const response = await Event.deleteOne({_id: req.params._id});
         res.json({message: 'Event deleted'});
